EMP-184: force an error to test PR can not be merged on error

+++ equinity-historical-data/src/main/java/uk/gov/justice/laa/crime/equinity/historicaldata/model/TaskImageFilesModel.java
@@ -14,7 +14,7 @@ import org.hibernate.annotations.Immutable;
 public class TaskImageFilesModel {
 
     @Id
-    private Integer ID;
+    private Long ID;
 
     @Column(name = "OFDImage")
     @Lob
